+++ custom-recipes/pi-system-retrieve-event-frames/recipe.py
@@ -4,7 +4,10 @@ import copy
 from dataiku.customrecipe import get_input_names_for_role, get_recipe_config, get_output_names_for_role
 import pandas as pd
 from safe_logger import SafeLogger
-from osisoft_plugin_common import get_credentials, get_interpolated_parameters, get_advanced_parameters, check_debug_mode, PerformanceTimer
+from osisoft_plugin_common import (
+    get_credentials, get_interpolated_parameters,
+    get_advanced_parameters, check_debug_mode, PerformanceTimer, get_max_count
+)
 from osisoft_constants import OSIsoftConstants
 from osisoft_client import OSIsoftClient
 
@@ -23,6 +26,7 @@ logger.info("Initialization with config={}".format(logger.filter_secrets(config)
 
 auth_type, username, password, server_url, is_ssl_check_disabled = get_credentials(config)
 is_debug_mode = check_debug_mode(config)
+max_count = get_max_count(config)
 
 use_server_url_column = config.get("use_server_url_column", False)
 if not server_url and not use_server_url_column:
@@ -69,7 +73,11 @@ with output_dataset.get_writer() as writer:
         event_frame_webid = input_parameters_row.get("WebId")
 
         if client is None or previous_server_url != server_url:
-            client = OSIsoftClient(server_url, auth_type, username, password, is_ssl_check_disabled=is_ssl_check_disabled, is_debug_mode=is_debug_mode, network_timer=network_timer)
+            client = OSIsoftClient(
+                server_url, auth_type, username, password,
+                is_ssl_check_disabled=is_ssl_check_disabled, is_debug_mode=is_debug_mode,
+                network_timer=network_timer
+            )
             previous_server_url = server_url
         object_id = input_parameters_row.get(path_column)
         item = None
@@ -84,14 +92,15 @@ with output_dataset.get_writer() as writer:
                 interval=interval,
                 sync_time=sync_time,
                 boundary_type=boundary_type,
-                can_raise=False
+                can_raise=False,
+                max_count=max_count
             )
         elif use_batch_mode:
             buffer.append(object_id)
             batch_buffer_size += 1
             if (batch_buffer_size >= batch_size) or (absolute_index == nb_rows_to_process):
                 rows = client.get_rows_from_webids(
-                    buffer, data_type,
+                    buffer, data_type, max_count,
                     can_raise=False,
                     batch_size=batch_size
                 )
@@ -108,6 +117,7 @@ with output_dataset.get_writer() as writer:
                 interval=interval,
                 sync_time=sync_time,
                 boundary_type=boundary_type,
+                max_count=max_count,
                 can_raise=False
             )
         unnested_items_rows = []