+++ custom-recipes/pi-system-retrieve-list/recipe.py
@@ -6,7 +6,7 @@ from safe_logger import SafeLogger
 from osisoft_plugin_common import (
     get_credentials, get_interpolated_parameters, normalize_af_path,
     get_combined_description, get_base_for_data_type, check_debug_mode,
-    PerformanceTimer
+    PerformanceTimer, get_max_count
 )
 from osisoft_client import OSIsoftClient
 from osisoft_constants import OSIsoftConstants
@@ -26,6 +26,7 @@ logger.info("Initialization with config config={}".format(logger.filter_secrets(
 
 auth_type, username, password, server_url, is_ssl_check_disabled = get_credentials(config)
 is_debug_mode = check_debug_mode(config)
+max_count = get_max_count(config)
 
 use_server_url_column = config.get("use_server_url_column", False)
 if not server_url and not use_server_url_column:
@@ -85,6 +86,7 @@ with output_dataset.get_writer() as writer:
                 interval=interval,
                 sync_time=sync_time,
                 boundary_type=boundary_type,
+                max_count=max_count,
                 can_raise=False,
                 object_id=object_id
             )
@@ -97,6 +99,7 @@ with output_dataset.get_writer() as writer:
                 interval=interval,
                 sync_time=sync_time,
                 boundary_type=boundary_type,
+                max_count=max_count,
                 can_raise=False,
                 endpoint_type="AF"
             )