+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fix Toolbox issue with tags and paths containing spaces
 - Fix Toolbox result consistency when using web IDs.
 - Add option to change max count value
+- Add performance timers
 
 ## [Version 1.0.1](https://github.com/dataiku/dss-plugin-pi-server/releases/tag/v1.0.1) - Feature release - 2023-01-05